chore(bundler): migrate bundler configuration across packages
