+++ configs/tsdown-config/README.md
@@ -1,0 +1,1 @@
+# `@aura-stack/tsdown-config`
