+++ configs/tsdown-config/tsdown.config.ts
@@ -1,0 +1,15 @@
+import type { UserConfig } from "tsdown"
+
+/**
+ * Tsdown base configuration used by Aura Auth packages
+ */
+export const tsdownConfig: UserConfig = {
+    entry: ["src"],
+    format: ["esm", "cjs"],
+    dts: true,
+    clean: true,
+    fixedExtension: false,
+    deps: {
+        onlyBundle: false,
+    },
+}
