--- configs/tsup-config/README.md
@@ -1,1 +1,0 @@
-# `@aura-auth/tsup-config`
