+++ package.json
@@ -43,7 +43,6 @@
     "oxfmt": "^0.41.0",
     "oxlint": "^1.56.0",
     "tsdown": "^0.21.8",
-    "tsup": "^8.5.1",
     "turbo": "^2.7.5",
     "typescript": "catalog:typescript",
     "vitest": "catalog:vitest"
