+++ packages/core/package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "description": "Core auth for @aura-stack/auth",
   "scripts": {
-    "dev": "tsup --watch",
-    "build": "tsup",
+    "dev": "tsdown --watch",
+    "build": "tsdown",
     "lint": "oxlint",
     "lint:fix": "oxlint --fix",
     "test": "vitest --run",
@@ -91,7 +91,7 @@
   },
   "devDependencies": {
     "@aura-stack/tsconfig": "workspace:*",
-    "@aura-stack/tsup-config": "workspace:*",
+    "@aura-stack/tsdown-config": "workspace:*",
     "typescript": "catalog:typescript",
     "vitest": "catalog:vitest"
   }
