--- packages/express/tsup.config.ts
@@ -1,4 +1,0 @@
-import { tsupConfig } from "@aura-stack/tsup-config"
-import { defineConfig } from "tsup"
-
-export default defineConfig(tsupConfig)
