+++ packages/hono/package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "description": "Hono integration for Aura Stack authentication library",
   "scripts": {
-    "dev": "tsup --watch",
-    "build": "tsup",
+    "dev": "tsdown --watch",
+    "build": "tsdown",
     "lint": "oxlint",
     "lint:fix": "oxlint --fix",
     "test": "vitest --run",
@@ -66,7 +66,7 @@
   },
   "devDependencies": {
     "@aura-stack/tsconfig": "workspace:*",
-    "@aura-stack/tsup-config": "workspace:*",
+    "@aura-stack/tsdown-config": "workspace:*",
     "hono": "catalog:hono",
     "@types/bun": "catalog:bun",
     "vitest": "catalog:vitest"
