--- packages/hono/tsup.config.ts
@@ -1,4 +1,0 @@
-import { defineConfig } from "tsup"
-import { tsupConfig } from "@aura-stack/tsup-config"
-
-export default defineConfig(tsupConfig)
