+++ packages/jose/package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "description": "JOSE utilities for @aura-stack/auth",
   "scripts": {
-    "dev": "tsup --watch",
-    "build": "tsup",
+    "dev": "tsdown --watch",
+    "build": "tsdown",
     "lint": "oxlint",
     "lint:fix": "oxlint --fix",
     "test": "vitest --run",
@@ -78,7 +78,7 @@
   },
   "devDependencies": {
     "@aura-stack/tsconfig": "workspace:*",
-    "@aura-stack/tsup-config": "workspace:*",
+    "@aura-stack/tsdown-config": "workspace:*",
     "typescript": "catalog:typescript",
     "vitest": "catalog:vitest"
   }
