+++ packages/jose/tsconfig.json
@@ -7,6 +7,6 @@
     "allowImportingTsExtensions": true,
     "noEmit": true
   },
-  "include": ["src", "tsup.config.ts", "test"],
+  "include": ["src", "tsdown.config.ts", "test"],
   "exclude": ["dist", "node_modules"]
 }
