+++ packages/next/package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "description": "Next.js integration for Aura Stack authentication library",
   "scripts": {
-    "dev": "tsup --watch",
-    "build": "tsup",
+    "dev": "tsdown --watch",
+    "build": "tsdown",
     "lint": "oxlint",
     "lint:fix": "oxlint --fix",
     "format": "oxfmt",
@@ -69,7 +69,7 @@
   },
   "devDependencies": {
     "@aura-stack/tsconfig": "workspace:*",
-    "@aura-stack/tsup-config": "workspace:*",
+    "@aura-stack/tsdown-config": "workspace:*",
     "next": "catalog:next",
     "react": "catalog:react",
     "react-dom": "catalog:react"
