+++ packages/rate-limiter/package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "description": "Rate limiter for @aura-stack/auth",
   "scripts": {
-    "dev": "tsup --watch",
-    "build": "tsup",
+    "dev": "tsdown --watch",
+    "build": "tsdown",
     "lint": "oxlint",
     "lint:fix": "oxlint --fix",
     "test": "vitest --run",
@@ -54,7 +54,7 @@
   "license": "MIT",
   "devDependencies": {
     "@aura-stack/tsconfig": "workspace:*",
-    "@aura-stack/tsup-config": "workspace:*",
+    "@aura-stack/tsdown-config": "workspace:*",
     "vitest": "catalog:vitest"
   },
   "packageManager": "pnpm@10.15.0"
