+++ packages/rate-limiter/tsdown.config.ts
@@ -1,0 +1,4 @@
+import { tsdownConfig } from "@aura-stack/tsdown-config"
+import { defineConfig, type UserConfig } from "tsdown"
+
+export default defineConfig(tsdownConfig as UserConfig)
