--- packages/react-router/tsup.config.ts
@@ -1,7 +1,0 @@
-import { tsupConfig } from "@aura-stack/tsup-config"
-import { defineConfig } from "tsup"
-
-export default defineConfig({
-    ...tsupConfig,
-    external: ["react", "react-dom", "react-router"],
-})
